ref: ざっくりUsecaseとRepositoryを分けた
